requirements.txt
keras>=3.0.0
torch
torchvision
gradio
numpy
matplotlib
seaborn
scikit-learn
pillow
python-dotenv
